Reorganize files since they're shared between projects, organization, and lab
