+++ app/lib/nav-helpers/userHasLabAccess.js
