+++ app/pages/project/components/ProjectNavbar/components/NavLink/NavLink.jsx
@@ -1,10 +1,9 @@
-import _ from 'lodash';
 import { Link } from 'react-router';
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { pxToRem } from '../../../../../../theme';
-import ExternalLink from '../../../ExternalLinksBlock/components/ExternalLink';
+import ExternalLink from '../../../../../../components/ExternalLinksBlock/components/ExternalLink';
 
 const commonStyles = `
   display: block;
