+++ test/helperFunctions.js
